Relocated Mock Input to header and Moved CV state to the top
